+++ glmdenoise/r2_nom_denom.py
@@ -1,0 +1,18 @@
+import numpy as np
+
+
+def R2_nom_denom(y, yhat):
+    """ Calculates the nominator and denomitor for calculating R-squared
+
+    Args:
+        y (array): data
+        yhat (array): predicted data data
+
+    Returns:
+        nominator (float or array), denominator (float or array)
+    """
+    y, yhat = np.array(y), np.array(yhat)
+    with np.errstate(divide="ignore", invalid="ignore"):
+        nom = np.sum((y - yhat) ** 2, axis=0)
+        denom = np.sum(y ** 2, axis=0)  # Kendricks denominator
+    return nom, denom
